+++ searchlib/src/vespa/searchlib/datastore/free_list_raw_allocator.hpp
@@ -25,7 +25,7 @@ FreeListRawAllocator<EntryT, RefT>::alloc(size_t numElems)
     assert(state.getArraySize() == numElems);
     RefT ref = state.popFreeList();
     // If entry ref is not aligned we must scale the offset according to array size as it was divided when the entry ref was created.
-    uint64_t offset = !RefT::isAlignedType ? ref.offset() * state.getArraySize() : ref.offset();
+    size_t offset = !RefT::isAlignedType ? ref.offset() * state.getArraySize() : ref.offset();
     EntryT *entry = _store.template getBufferEntry<EntryT>(ref.bufferId(), offset);
     return HandleType(ref, entry);
 }
